// apps/SimpleCodeApp/src/generated/services/RetrieveAuditDetailsService.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

import type { IOperationResult } from '@microsoft/power-apps/data';
import { dataSourcesInfo } from '../../../.power/schemas/appschemas/dataSourcesInfo';
import { getClient } from '@microsoft/power-apps/data';

export class RetrieveAuditDetailsService {
  private static readonly dataSourceName = 'retrieveauditdetails';

  private static readonly client = getClient(dataSourcesInfo);

  /**
   * @param id type: `Guid`
   */
  public static async RetrieveAuditDetails(id: string): Promise<IOperationResult<Record<string, unknown>>> {
    const params: { id: string } = { id };
    const result = await RetrieveAuditDetailsService.client.executeAsync<{ id: string }, Record<string, unknown>>(
      {
        dataverseRequest: {
          action: 'customapi',
          parameters: {
            operationName: 'RetrieveAuditDetails',
            tableName: RetrieveAuditDetailsService.dataSourceName,
            body: params
          }
        },
      });
    return result;
  }
}